[dev] Revert "feature: [dev-8574] Advanced Editor updates (#1372)"
This reverts commit 56d002e312f836c0831629f5efaa6ce2ac9863d5.
