+++ packages/chart/src/components/EditorPanel/EditorPanel.tsx
@@ -2708,7 +2708,7 @@ const EditorPanel = () => {
             <Panels.Annotate name='Text Annotations' />
             {/* {(config.visualizationType === 'Bar' || config.visualizationType === 'Line') && <Panels.DateHighlighting name='Date Highlighting' />} */}
           </Accordion>
-          {config.type !== 'Spark Line' && <AdvancedEditor loadConfig={updateConfig} config={config} convertStateToConfig={convertStateToConfig} />}
+          {config.type !== 'Spark Line' && <AdvancedEditor loadConfig={updateConfig} state={config} convertStateToConfig={convertStateToConfig} />}
         </Layout.Sidebar>
       </ErrorBoundary>
     </EditorPanelContext.Provider>
