+++ packages/core/components/AdvancedEditor.jsx
@@ -1,33 +1,23 @@
 import React, { useState, useEffect } from 'react'
-import MapIcon from '../../assets/map-folded.svg'
-import ChartIcon from '../../assets/icon-chart-bar.svg'
-import MarkupIncludeIcon from '../../assets/icon-code.svg'
-import { JsonEditor, UpdateFunction } from 'json-edit-react'
-import { formatConfigBeforeSave as stripConfig } from '../../helpers/formatConfigBeforeSave'
-import './advanced-editor-styles.css'
-import _ from 'lodash'
+import MapIcon from '../assets/map-folded.svg'
+import ChartIcon from '../assets/icon-chart-bar.svg'
+import MarkupIncludeIcon from '../assets/icon-code.svg'
 
-export const AdvancedEditor = ({ loadConfig, config, convertStateToConfig, onExpandCollapse = () => {} }) => {
-  const [advancedToggle, _setAdvancedToggle] = useState(false)
-  const [configTextboxValue, setConfigTextbox] = useState('')
-  const setAdvancedToggle = val => {
-    _setAdvancedToggle(val)
-    onExpandCollapse()
-  }
-
-  const onUpdate: UpdateFunction = val => {
-    setConfigTextbox(JSON.stringify(val.newData))
-  }
+export const AdvancedEditor = ({ loadConfig, state, convertStateToConfig }) => {
+  const [advancedToggle, setAdvancedToggle] = useState(false)
+  const [configTextboxValue, setConfigTextbox] = useState({})
 
   useEffect(() => {
-    let parsedConfig = stripConfig(config)
-    if (config.type !== 'dashboard') {
-      parsedConfig = convertStateToConfig()
+    let parsedData = state
+    if (state.type !== 'dashboard') {
+      parsedData = convertStateToConfig()
+    } else {
+      parsedData = JSON.parse(JSON.stringify(parsedData))
     }
-    const formattedConfig = JSON.stringify(parsedConfig, undefined, 2)
+    const formattedData = JSON.stringify(parsedData, undefined, 2)
 
-    setConfigTextbox(formattedConfig)
-  }, [config])
+    setConfigTextbox(formattedData)
+  }, [state])
 
   const typeLookup = {
     chart: ['Charts', 'https://www.cdc.gov/wcms/4.0/cdc-wp/data-presentation/bar-chart.html', <ChartIcon />],
@@ -36,13 +26,13 @@ export const AdvancedEditor = ({ loadConfig, config, convertStateToConfig, onExp
     'markup-include': ['Markup Include', 'https://www.cdc.gov/wcms/4.0/cdc-wp/data-presentation/Markup-Include.html', <MarkupIncludeIcon />]
   }
 
-  if (!config.type) return <></>
+  if (!state.type) return <></>
   return (
     <>
-      <a href={typeLookup[config.type][1]} target='_blank' rel='noopener noreferrer' className='guidance-link'>
-        {typeLookup[config.type][2]}
+      <a href={typeLookup[state.type][1]} target='_blank' rel='noopener noreferrer' className='guidance-link'>
+        {typeLookup[state.type][2]}
         <div>
-          <span className='heading-3'>Get Help with {typeLookup[config.type][0]}</span>
+          <span className='heading-3'>Get Help with {typeLookup[state.type][0]}</span>
           <p>Examples and documentation</p>
         </div>
       </a>
@@ -62,21 +52,16 @@ export const AdvancedEditor = ({ loadConfig, config, convertStateToConfig, onExp
               This tool displays the actual <acronym title='JavaScript Object Notation'>JSON</acronym> configuration that is generated by this editor and allows you to edit properties directly and apply them.
             </p>
             <button
-              className='btn '
+              className='btn'
               onClick={() => {
                 navigator.clipboard.writeText(configTextboxValue)
+                alert('Copied!')
               }}
             >
               Copy to Clipboard
             </button>
-            <JsonEditor className='advanced-json-editor' data={JSON.parse(configTextboxValue)} onUpdate={onUpdate} rootName='' />
-            <button
-              className='btn full-width'
-              onClick={() => {
-                loadConfig(JSON.parse(configTextboxValue))
-                setAdvancedToggle(!advancedToggle)
-              }}
-            >
+            <textarea value={configTextboxValue} onChange={event => setConfigTextbox(event.target.value)} />
+            <button className='btn full-width' onClick={() => loadConfig(JSON.parse(configTextboxValue))}>
               Apply
             </button>
           </React.Fragment>
