--- packages/core/components/AdvancedEditor/advanced-editor-styles.css
@@ -1,3 +1,0 @@
-.advanced-json-editor {
-  width: 100vw;
-}
