--- packages/core/components/AdvancedEditor/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AdvancedEditor'
