+++ packages/core/package.json
@@ -21,7 +21,6 @@
   "license": "Apache-2.0",
   "dependencies": {
     "html2canvas": "^1.4.1",
-    "json-edit-react": "^1.12.0",
     "papaparse": "^5.3.0",
     "prop-types": "^15.8.1",
     "react-accessible-accordion": "^5.0.0",
