--- packages/dashboard/src/components/VisualizationsPanel/index.ts
@@ -1,1 +1,0 @@
-export { default } from './VisualizationsPanel'
