--- packages/dashboard/src/components/VisualizationsPanel/visualizations-panel-styles.css
@@ -1,12 +1,0 @@
-.visualizations-panel {
-  background-color: #fff;
-  padding: 1em;
-  width: var(--editorWidth);
-  border-right: #c7c7c7 1px solid;
-  z-index: 1;
-  overflow-y: scroll;
-
-  &.advanced-editor {
-    width: 50vw;
-  }
-}
