+++ packages/dashboard/src/scss/variables.scss
@@ -1,5 +1,1 @@
 $editorWidth: 350px;
-
-:root {
-  --editorWidth: #{$editorWidth};
-}
