+++ packages/map/src/components/EditorPanel/components/EditorPanel.tsx
@@ -3057,7 +3057,7 @@ const EditorPanel = ({ columnsRequiredChecker }) => {
           {state.general.geoType === 'us' && <Panels.PatternSettings name='Pattern Settings' />}
           {state.general.geoType !== 'us-county' && <Panels.Annotate name='Text Annotations' />}
         </Accordion>
-        <AdvancedEditor loadConfig={loadConfig} config={state} convertStateToConfig={convertStateToConfig} />
+        <AdvancedEditor loadConfig={loadConfig} state={state} convertStateToConfig={convertStateToConfig} />
       </Layout.Sidebar>
     </ErrorBoundary>
   )
